Move server spi package to server event package
