+++ platform/dolphin-platform-distributed-eventbus/src/main/resources/META-INF/services/com.canoo.platform.remoting.server.com.canoo.platform.remoting.server.event.spi.EventBusProvider
