+++ platform/dolphin-platform-remoting-server/src/main/java/com/canoo/dp/impl/server/event/DefaultEventBusProvider.java
@@ -17,7 +17,7 @@ package com.canoo.dp.impl.server.event;
 
 import com.canoo.dp.impl.server.config.RemotingConfiguration;
 import com.canoo.platform.remoting.server.event.RemotingEventBus;
-import com.canoo.platform.remoting.server.spi.EventBusProvider;
+import com.canoo.platform.remoting.server.event.spi.EventBusProvider;
 import org.apiguardian.api.API;
 
 import static org.apiguardian.api.API.Status.INTERNAL;
